Add ability to ignore unparseable HCL files
As we increase usage of HCL across our codebases, we've run into a few
instances of files that are valid HCL (which are consumed just fine by
Consul, for example), but cannot (for whatever reason) be recognized
as valid HCL files by `packer fmt` and `packer lint`, which forms the
basis for the formatter and linter in this repository.

Rather than contorting our formatting and linting process in
these repositories, we can now specify an `--ignore` flag for our
formatter, allowing us to single out these offending files and, well,
ignore them.

As such, the main change here is in the `docker_entrypoint.sh`
script. Since this repository hasn't received any action since the
initial commit, however, we'll take the opportunity to upgrade the
build process to be more in line with our other
repositories.

Specifically:

- We use Pants to get easy access to `shellcheck`, `shfmt` and
  `hadolint`.
- We start using the code within this repository to lint and
  format the HCL here (rather than relying on published images; this
  serves as a very basic smoke test).
- We make use of Docker Compose v2, rather than v1.
- We spruce up the Makefile to mirror the ones in our other
  codebases.
- Finally, we upgrade the Packer image we base everything on to the
  current version `1.8.2`

Signed-off-by: Christopher Maier <chris@graplsecurity.com>